WebXR: Fix DOM overlay for out of process iframes (OOPIF)
Propagate the "is_xr_overlay" fullscreen request type flag for OOPIF
requests via mojo, and ensure that visibility is set correctly for
out of process container iframe elements.

Note for testing: Currently, Chrome on Android only enforces site
isolation for sites with login active, and not on low memory devices.
Run with the --site-per-process command line flag to override this.

Bug: 1101193
Change-Id: Ic8736e90f28fbda2d6cb1f665cbd0804c3f38cc3